moved new Camera Class to base_classes folder
updated all the references in code
