+++ config/config_base.yaml
@@ -1,7 +1,7 @@
 #- Define paths and base settings
 # (they will be converted to pathlib.Path when parsing this yaml file)
 paths:
-  image_dir: "data/img2022"
+  image_dir: "data/img2022" # "data/img2021" #
   calibration_dir: "data/calib"
   results_dir: "res"
   camera_names: ["p1", "p2"]
@@ -37,7 +37,7 @@ georef:
 
   #- Targets' information
   # Target information must be stored in file one file per image saved with the name of the relative image (with target_file_ext extension)
-  target_dir: "data/targets"
+  target_dir: "data/targets" # "data/targets2021" #
   target_file_ext: ".csv"
 
 #- Other On-Off switches
@@ -98,10 +98,3 @@ tracking:
   do_viz_tile: false
   rowDivisor: 2
   colDivisor: 4
-#- Tile processing
-# useTile: True
-# writeTile2Disk: False
-# do_viz_tile: True
-# rowDivisor: 2
-# colDivisor: 4
-# overlap: 400
