+++ data/targets2021/IMG_0520.csv
@@ -1,0 +1,3 @@
+label,x,y
+F2,1397.7201,3713.5696
+F4,4189.4971,3017.1682
