+++ data/targets2021/IMG_0523.csv
@@ -1,0 +1,3 @@
+label,x,y
+F2,1398.8444,3713.2534
+F4,4190.0669,3017.6262
