+++ data/targets2021/IMG_0526.csv
@@ -1,0 +1,3 @@
+label,x,y
+F2,1399.8260,3713.8630
+F4,4191.2261,3018.4219
