+++ data/targets2021/IMG_0528.csv
@@ -1,0 +1,3 @@
+label,x,y
+F2,1400.1013,3714.2766
+F4,4190.9790,3018.8335
