+++ data/targets2021/IMG_2131.csv
@@ -1,0 +1,3 @@
+label,x,y
+F2,3453.2603,3650.5872
+F4,5260.6113,2456.3042
