+++ data/targets2021/IMG_2134.csv
@@ -1,0 +1,3 @@
+label,x,y
+F2,3453.4663,3650.4390
+F4,5260.3042,2456.3618
