+++ data/targets2021/IMG_2137.csv
@@ -1,0 +1,3 @@
+label,x,y
+F2,3454.3340,3649.4006
+F4,5261.1826,2455.7463
