+++ data/targets2021/IMG_2139.csv
@@ -1,0 +1,3 @@
+label,x,y
+F2,3454.5378,3649.2986
+F4,5261.0034,2456.0439
