+++ data/targets2021/target_world_p1.csv
@@ -1,0 +1,15 @@
+label,X,Y,Z
+T2,-499.8550,402.0301,240.3745
+T3,-302.8139,442.8938,221.9927
+T4,-543.4710,365.4121,217.7905
+F2,49.6488,192.0875,71.7466
+F3,-15.5549,315.3171,117.3462
+F4,41.5596,274.7651,92.0252
+F5,53.2401,163.0449,78.2960
+11,-194.7426,287.7045,150.1151
+12,-47.1266,257.7534,119.0749
+13,-21.9544,227.5956,95.2070
+15,-19.5291,238.1006,96.6108
+16,8.2768,244.6750,80.6632
+17,-6.8296,285.0130,98.5180
+19,-141.9356,324.2512,141.5215
