+++ lib/base_classes/images.py
