+++ lib/visualization.py
@@ -33,8 +33,7 @@ import matplotlib
 from typing import List, Union
 from pathlib import Path
 
-from lib.classes import CameraNew as Camera
-
+from lib.base_classes.camera import Camera
 from lib.classes import Features
 from lib.utils.utils import compute_reprojection_error
 from lib.geometry import project_points
